fix: harden GHE URL detection and extract shared GitHub helpers
